[app] Remove app package and put in app.go
- no more interface, just a struct
