+++ _example/github/cmd/icehubctl/main.go
@@ -4,17 +4,17 @@ import (
 	"fmt"
 	"os"
 
-	"github.com/at15/go.ice/ice/app"
+	"github.com/at15/go.ice/ice"
 
 	"github.com/at15/go.ice/_example/github/pkg/common"
 )
 
 func main() {
-	a := app.New(
-		app.Name("icehubctl"),
-		app.Description("Client of IceHub, which is an example GitHub integration service using go.ice"),
-		app.Version(common.Version()))
-	root := app.NewCmd(a)
+	app := ice.New(
+		ice.Name("icehubctl"),
+		ice.Description("Client of IceHub, which is an example GitHub integration service using go.ice"),
+		ice.Version(common.Version()))
+	root := ice.NewCmd(app)
 	if err := root.Execute(); err != nil {
 		fmt.Fprintln(os.Stderr, err)
 		os.Exit(1)
