--- ice/app/client.go
@@ -1,4 +1,0 @@
-package app
-
-type Client struct {
-}
