--- ice/app/server.go
@@ -1,4 +1,0 @@
-package app
-
-type Server struct {
-}
